Combine extract-meta and TypeScript generator.
