+++ @plotly/dash-generator-test-component-typescript/babel.config.js
@@ -1,0 +1,8 @@
+const presets = [
+    '@babel/preset-env',
+    '@babel/preset-react'
+];
+
+const plugins = [];
+
+module.exports = { presets, plugins };
