+++ @plotly/dash-generator-test-component-typescript/package.json
@@ -10,19 +10,23 @@
   "scripts": {
     "build:js": "webpack --mode production",
     "setup": "python setup.py sdist",
-    "build:py_and_r": "dash-generate-components ./src/components dash_generator_test_component_typescript --ts && cp base/** dash_generator_test_component_typescript/ && dash-generate-components ./src/components dash_generator_test_component_typescript --r-prefix 'dgtc_ts' --ts",
+    "build:py_and_r": "dash-generate-components ./src/components dash_generator_test_component_typescript && cp base/** dash_generator_test_component_typescript/ && dash-generate-components ./src/components dash_generator_test_component_typescript --r-prefix 'dgtc_ts'",
     "build": "run-s build:js build:py_and_r setup",
     "test": "jest"
   },
   "author": "Chris Parmer <chris@plotly.com>",
   "license": "MIT",
   "devDependencies": {
+    "@babel/preset-env": "^7.16.11",
+    "@babel/preset-react": "^7.16.7",
     "@types/jest": "^27.4.1",
     "@types/react": "^17.0.39",
+    "babel-loader": "^8.2.4",
     "jest": "^27.5.1",
     "npm-run-all": "^4.1.5",
     "ramda": "^0.28.0",
     "react": "16.13.0",
+    "react-docgen": "^5.4.0",
     "react-dom": "16.13.0",
     "ts-jest": "^27.1.3",
     "ts-loader": "^8.2.3",
