+++ @plotly/dash-generator-test-component-typescript/src/index.ts
@@ -1,5 +1,11 @@
 import TypeScriptComponent from './components/TypeScriptComponent';
 import TypeScriptClassComponent from './components/TypeScriptClassComponent';
 import MemoTypeScriptComponent from './components/MemoTypeScriptComponent';
+import StandardComponent from './components/StandardComponent.react';
 
-export {TypeScriptComponent, TypeScriptClassComponent, MemoTypeScriptComponent};
+export {
+    TypeScriptComponent,
+    TypeScriptClassComponent,
+    MemoTypeScriptComponent,
+    StandardComponent
+};
