+++ dash/development/component_generator.py
@@ -48,16 +48,13 @@ def generate_components(
     jlprefix=None,
     metadata=None,
     keep_prop_order=None,
-    typescript=False,
 ):
 
     project_shortname = project_shortname.replace("-", "_").rstrip("/\\")
 
     is_windows = sys.platform == "win32"
 
-    extract_path = pkg_resources.resource_filename(
-        "dash", "meta-ts.js" if typescript else "extract-meta.js"
-    )
+    extract_path = pkg_resources.resource_filename("dash", "extract-meta.js")
 
     reserved_patterns = "|".join(f"^{p}$" for p in reserved_words)
 
@@ -70,9 +67,8 @@ def generate_components(
     if not metadata:
         env = os.environ.copy()
 
-        if typescript:
-            # Ensure local node modules is used when the script is packaged.
-            env["MODULES_PATH"] = os.path.abspath("./node_modules")
+        # Ensure local node modules is used when the script is packaged.
+        env["MODULES_PATH"] = os.path.abspath("./node_modules")
 
         cmd = shlex.split(
             f'node {extract_path} "{ignore}" "{reserved_patterns}" {components_source}',
@@ -225,11 +221,6 @@ def component_build_arg_parser():
         "props. Pass the 'ALL' keyword to have every component retain "
         "its original prop order.",
     )
-    parser.add_argument(
-        "--ts",
-        action="store_true",
-        help="Generate components from typescript source. (.tsx files)",
-    )
     return parser
 
 
@@ -246,7 +237,6 @@ def cli():
         rsuggests=args.r_suggests,
         jlprefix=args.jl_prefix,
         keep_prop_order=args.keep_prop_order,
-        typescript=args.ts,
     )
